backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
astunparse==1.6.3
autopep8==2.3.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
flake8==7.1.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
openai==1.35.13
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
protobuf==4.25.3
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
replicate==0.28.0
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
